(PE-1927) Refactor tasks integration test
Improve the `get-all-tasks` test utility by simplifying the url
construction and pulling out the HTTP request options into a var so they
can be reused.

Add a `sort-tasks` utility to make the `all-tasks-with-env` test pass
consistently, and pull the config used to start puppetserver out into
its own var so it can be reused.