(maint) Clarify purposes of jruby tasks utils
Add an explanation to the end of the docstrings of the task-generating
utilities in puppetlabs.services.jruby.tasks-test that explains that
these utilities take the place to generate the task as an input, rather
than generating them from a fixed starting point (as the task-generating
utilities in the main testutils namespace do). Also add `empty` to their
names to indicate that the files they create are all empty (either
literally blank, or containing empty JSON).

Pull out the config generation logic from the `all-tasks-test` into
a function called `puppet-config`, so it can be reused in the
forthcoming `task-info-test`.

Break out the `env-dir` function from the `all-tasks-test` into
a top-level function that also takes the code dir, so it can be reused
as well.